// v2/lib/types.ts
export interface MunicipioRawData {
    ano: number;
    sigla_uf: string;
    sigla_uf_nome: string;
    id_municipio: number;
    municipio: string;
    cidade: string;
    pop: number;
    IDHM: number;
    "IDHM Renda": number;
    "IDHM Educação": number;
    "IDHM Longevidade": number;
    total_docentes: number;
    total_alunos: number;
    quantidade_profissionais_enfermeiro: number;
    quantidade_profissionais_medico: number;
    valor_total_ato_profissional: number;
    numero_de_empresas_agronegocio: number;
    quantidade_vinculos_ativos_agronegocio: number;
    numero_de_empresas_comercio: number;
    quantidade_vinculos_ativos_comercio: number;
    numero_de_empresas_construcao: number;
    quantidade_vinculos_ativos_construcao: number;
    numero_de_empresas_educacao: number;
    quantidade_vinculos_ativos_educacao: number;
    numero_de_empresas_industria: number;
    quantidade_vinculos_ativos_industria: number;
    numero_de_empresas_saude: number;
    quantidade_vinculos_ativos_saude: number;
    numero_de_empresas_turismo: number;
    quantidade_vinculos_ativos_turismo: number;
}

export interface MunicipioRow {
    ano: number | string | null;
    siglaUf: string | null;
    siglaUfNome: string | null;
    idMunicipio: number | null;
    municipio: string | null;
    cidade: string | null;
    pop: number | null;
    idhm: number | null;
    idhmRenda: number | null;
    idhmEducacao: number | null;
    idhmLongevidade: number | null;
    idhmMulheres: number | null;
    idhmHomens: number | null;
    idhmNegros: number | null;
    idhmBrancos: number | null;
    totalDocentes: number | null;
    totalAlunos: number | null;
    taxaAprovacao: number | null;
    indicadorRendimento: number | null;
    ideb: number | null;
    quantidadeProfissionaisEnfermeiro: number | null;
    quantidadeProfissionaisMedico: number | null;
    valorTotalAtoProfissional: number | null;
    numeroDeEmpresasAgronegocio: number | null;
    quantidadeVinculosAtivosAgronegocio: number | null;
    numeroDeEmpresasComercio: number | null;
    quantidadeVinculosAtivosComercio: number | null;
    massaSalarialComercio: number | null;
    numeroDeEmpresasConstrucao: number | null;
    quantidadeVinculosAtivosConstrucao: number | null;
    massaSalarialConstrucao: number | null;
    numeroDeEmpresasEducacao: number | null;
    quantidadeVinculosAtivosEducacao: number | null;
    massaSalarialEducacao: number | null;
    numeroDeEmpresasIndustria: number | null;
    quantidadeVinculosAtivosIndustria: number | null;
    massaSalarialIndustria: number | null;
    numeroDeEmpresasSaude: number | null;
    quantidadeVinculosAtivosSaude: number | null;
    massaSalarialSaude: number | null;
    numeroDeEmpresasTurismo: number | null;
    quantidadeVinculosAtivosTurismo: number | null;
    massaSalarialTurismo: number | null;
}